Extract uniqBy into utils
