+++ src/hooks/usePagination.js
@@ -1,22 +1,7 @@
 import React from 'react';
 import { client } from 'client';
 import appConfig from 'app.config';
-
-export const uniqBy = (arr, pred) => {
-  const cb = typeof pred === 'function' ? pred : (o) => o[pred];
-
-  return [
-    ...arr
-      .reduce((map, item) => {
-        const key = item === null || item === undefined ? item : cb(item);
-
-        map.has(key) || map.set(key, item);
-
-        return map;
-      }, new Map())
-      .values(),
-  ];
-};
+import { uniqBy } from 'utils';
 
 /**
  * usePagination hook is focused on performing paginated queries using a custom query function.
