+++ src/utils/index.js
@@ -1,0 +1,3 @@
+import uniqBy from './uniqBy';
+
+export { uniqBy };
